Refactor reboot system
The system gets rebooted via kexec or if failed by a hard reboot.
For kexec it's required to load kernel, initrd and boot parameters
from the migrated system. Thus the kexec load must be done before
umount of the system and the actual reboot after umount of the
system. Therefore this patch splits the kernel-reboot service
into two services, whereas the umount service will be called
in between of them.