Added zypper migration console logging
If the migration process starts the console should show a
progress information. We use a dialog tailbox which shows
the output of the zypper migration plugin while it runs.
The output is processed using a systemd service connected
to the console. In addition the log file

    var/log/zypper_migrate.log

Is created inside of the system which gets migrated. This
allows for later inspection of what the migration plugin
did. This Fixes #27