Fix log file initialization
We agreed on writing the log file into the root filesystem
of the system to migrate. This implies that the first time
to initialize this log file is in the prepare service
after the mount system service has succeeded. Calling
set_logfile in the mount system service before the system
got mounted is the wrong place. This patch fixes this
and also improves the unit test for this condition.