Mount kernel file systems inside /system-root
The file systems were previously mounted in grub
service but because the migration service needs to
have access proc, dev and sys inside system-root
(e.g. when updating the bootloader) they are mounted
in an earlier step. There is no need to mount them
again in grub. They are unmounted properly before
rebooting.

This Fixes #32