Update prepare service on hosts file
In the cloud the translation from a name into an IP might also
be configured via the static etc/hosts file. In case of SUSE's
public cloud infrastructure the connected smt server is
configured that way at registration time. For the migration
process this means that this information must be present otherwise
the host to upgrade cannot be resolved.