Changed service type for console logger
Move to simple type service and set a restart policy.
The way the console logger was started before caused
a stop of the service as soon as the kernel or some
other program logs information on the console. However
during migration we want the console logger to be
actively occupy the console such that it's clear a
migration process is currently running