Handle distro_target
Long time ago zypper added a distro_target information into
the repository file and matches those with the product
information of the system. If the distro doesn't match
zypper refused to create the repository. In the migration
process using SMT this caused the zypper migration plugin
to fail because zypper never created repositories. That's
because in a migration process the repository and the
system to migrate never match. Therefore we bind mount
the /etc/products.d information from the migration live
system into the /system-root of the system to become
migrated. At the end of a successful migration the new
product information is copied to the migrated system