Revert/Refactor distro_target handling
zypper implements a handling for the distro_target attribute.
If present in the repository metadata, zypper compares the
value with the baseproduct <target> setup in /etc/products.d.
If they mismatch zypper refuses to create the repo. In the
process of a migration from distribution [A] to [B] this mismatch
always applies by design and prevents the zypper migration plugin
to work. In SCC or RMT the repositories doesn't contain the
distro_target attribute which is the reason why we don't see
this problem there. SMT however creates repos including
distro_target.

The refactored workaround solution is now to delete the target
specification in the baseproduct registration if present, because
the overlay mounting of etc/products.d did not work as it would
lead to a wrong upgrade path on response from SCC.

In addition a backup of the original products.d data is created
and used in the zypper migration plugin in case of an error