Add rootpart detection to grub activation script
The live migration image gets installed to the system again
because of the space limitation on /boot. This affects the
menuentry created on grub side in a way that we can't use
the pre-allocated pointer to the boot device but have to search
the root partition like in a real grub root entry. This patch
adds the needed code changes to locate the root part, insert
the needed filesystem module and initializes the root variable
to allow the loopback loading of the image. This Fixes #54