Support new cloud regionsrv client
The new cloud region service client comes with a plugin for
zypper that translates repo URIs into a specific plugin format
such that it's no longer possible to gain access to the SLES
repositories of our public cloud infrastructure by playing
tricks with snapshotting an on-demand instance without the
billingtag. However the new and fancy way makes migration of
such a system more difficult because the new repo URI format
must be known to zypper migration. This requires all software
components to be provided in the live migration image which
has been done in the image description but furthermore requires
that the cloud specific /etc/regionserverclnt.cfg and the
contents of /usr/lib/zypp/plugins from the system to migrate
are available in the migration live image. The later is
implemented in this commit and Fixes #69