Fix network setup service
The network setup service for the migration inherits the
network as it is configured on the system to become migrated.
As part of that process the network is restarted through
systemctl. However this is a problem because it interferes
with the network service dependency chain. This could result
in a failed network startup depending on how fast other
services e.g dbus are up and running. So the way we do it
causes a race condition. This commit deletes the network
startup from our network setup service. Our network setup
service runs by definition of the unit file before the
network.target. This means at the time the network service
activates the network our setup routine is already done and
a normal network startup should be guaranteed.