Split activation package from services
From a release process it doesn't make sense to keep the activation
script as a sub package of the services. The reason is because the
suse-migration-services package will be released to the migration
target distribution and the activation package will be released to
the migration start distribution. Thus the code is relevant on
different distributions and it is not useful to maintain that in
one package. This Fixes #74