Fix reboot race condition
The issue described here came up with the package split and
rename of the activation package. The grub setup service
deletes the activation package and calls grub2-mkconfig which
deletes the migration live boot entry as well as updates grub
to boot the migrated system. However if the activation package
could not be removed the process ends in a boot loop. Therefore
one part of this patch makes the removal of the migration
packages more robust. Another part provides detailed log
information about the removal and the grub setup in the log
file.

Last but not least the patch includes a refactoring of the
umount and reboot process which offers a race condition if
the grub-setup service has failed. The patch here also
Delete the umount service and put the code in the reboot
service. Only on reboot we umount the system. In case of
debugging we want to have full access to the root of the
system that should have become migrated. In addition the
systemd dependency chain for the kernel-load vs. reboot
service was racy and got fixed by a clear dependency
for grub-setup -> kernel-load -> reboot