Add support for zypper dup mode
By default the migration system uses the zypper migration plugin
to migrate the system. This works if the system is SLES registered
and an upgrade path from SCC exists. For system where this is not
possible an alternative method can be switched on with the
custom configuration option use_zypper_migration_plugin set
to: false. In this mode we just call zypper dup. Please note
this requires that the user has pre setup the repositories before
the migration system gets started. This Fixes #72