Inherit ssh host keys
The live image creates its ssh host key on boot. However that
host key does not match the host key from the system to migrate.
This leads to a warning message on a changed host key when a
ssh login attempt is made while the migration runs. As the
migration live system inherits the network setup it should
also inherit the ssh host key. This Fixes #86