Fixed SSH identity copy
The SSH host identity is provided by the server credentials that
makes up the fingerprint of that host. However the copy of the
files should exclude the encrypted private key ssh_host_key which
can only be read on the origin host. In fact a copy of that file
to another host will lead to a segfault in sshd and kills the
connection without any further information.