Allow to skip local fstab paths
If a device path is referenced as local path in the fstab file
but can't be found on the system, this device path is skipped.
Usually a device path references a storage node that exists
as /dev/.. node on the machine. However it's possible to also
reference files e.g for loop mount in the fstab file. If that
file reference is not present in the live migration system it
is ignored because considered harmless. file references in
fstab are sometimes used for swapfiles or other file mapped
data. For the purpose of the migration we consider such paths
as not belonging to the system. This assumption of course
could be wrong which is why we also write a warning message
to the log file once this situation is disovered. This
Fixes #110