Fixed network setup
The network setup was based on an overlay bind mount of the
entire /etc/sysconfig/network directory. However that directory
also contains script code and functions that are used by system
tools like netconfig. The scripts here are not compatible between
distributions. Thus the overmount of /etc/sysconfig/network/scripts
from SLE12 to SLE15 as one example breaks the netconfig tool and
prevents the update of the /etc/resolv.conf file on network
restart. Other negative effects of that overmounting are likely.
Therefore this commit changes the way we inherit the network setup.
Only the plugin directory and the interface configurations are
taken from the host, all the rest stays untouched.