Allow for vendor change during migration
The migration concept allows for the zypper dup as well as
for the zypper migrate approach. In the first case the user
is responsible for setting up the repos and we expect that
to be done in a way that vendor changes are acceptable. In
the second case a repository server is used to serve the
repos and we expect that registration instance to be trustworthy
such that potential vendor changes are also allowed