Add schema validation for migration config file
If the custom config file is empty or contains only comment lines,
it will now be ignored (rather than raising an exception and killing
the migration).  If the file is corrupt somehow, or violates schema
validation, an error will be logged and the migration will abort.

Validation is *also* applied to the default config file (the one
inside the migration image).  This makes automated testing easy,
because we'll effectively do automatic validation on any of the
test config files, whether they're for default config or custom
config.