Do kernel mounts in mount service (not in prepare service)
If the /dev, /proc and /sys mounts are done in the prepare service,
then exceptions thrown between mount system and prepare result in
the grub config not being applied.  Moving these mounts earlier in
the process ensures that the grub setup service has everything it
needs (see https://github.com/SUSE/suse-migration-services/pull/142
for more detailed discussion of the problem).