Prevent kernel filesystems from umount
Since the change to let systemd drive the reboot process
it is needed to keep the kernel filesystems proc, sys and dev
open to let systemd perform its job. The umounting prior to
reboot will be taken over by systemd as well. However if systemd
can't read e.g the state of kexec through /sys it can't
work as expected