Fixed registration check
The SUSEConnect call to check if the system is registered was
called in the live migration system environment. Of course there
it will fail because the live migration system is never registered.
The call has to be performed chrooted inside of the system we
want to migrate. Thus the caller arguments as well as the call
time in the preparation unit needs a fix