Conditionally check if the system is registered
Only if the zypper migration plugin is used it's required to
check if the system is registered. In case use_zypper_migration
is deactivated through the config file the migration will
run zypper dup with a customer specific set of repositories.
In this case it's not neccessarily required that this system
is registered to use a SUSE repository server.