Cleanup and fix logging
The logger class contained methods never used. In addition
the logging setup was wrong and only partially working. Any
unit file represents a systemd call. This means any unit file
is an extra program call which requires the initialization
of the python logging facility. The logging itself is also
extended in a way that any external program calls from the
Command class will now be logged and written to the logfile
at call time