Fixed handling of busy state on system-root
The console log service holds a busy state on the system-root
mount because the log file is written on the host to migrate.
At reboot time the service should be stopped to allow a clean
umount procedure prior reboot. In addition the console log
service should only restart on failure and not always