Respect host kernel boot options for migration
The kernel boot options used on the host to migrate can be
important for the migration live environment too. For example
if net.ifnames is passed is influences the network interface
names to become predictable. As the DMS inherits configuration
data from the host e.g the network setup, it's required that
also the kernel boot parameters matches.