Prevent use of shutil.copytree
The versions of copytree comes with a different set of
features depending on the used python interpreter. The
former implementation used an option which did not exist
on python in SLE15. To keep us safe from further surprises
I moved back to the simple but stable shutil.copy method.
We can come up with a refactoring PR when needed but not
combined with the fix for the certificates as this was the
original intention of the change