Cracefully handle cert copy exceptions
If the copy process of a certificate failed this should not
cause the entire migration process to stop. This commit
handles copy errors from the cert chain and turns them into
a log message. If in the subsequent chain of tasks the
migration failed because of missing certificates we will
see that in the log and in the attempt to access the repos
which is better than the python stacktrace on an unhandled
FileNotFoundError exception. Retated to Issue #197