Update option settings in zypper dup mode
When using the DMS in zypper dup mode the option
settings --allow-downgrade and --no-confirm are handy.
Even though I think with the global non-interactive
option set it should not be required to set no-confirm
in addition but the worst are questions being asked
so better add it. With allow-downgrade set the ability
to downgrade packages is given. With zypper dup the
repo setup is user defined custom set and therefore
downgrades should not be off the table