Setup package conflicts properly
migration services and migration activation conflicts.
The activation is installed on the host to migrate, the
services are installed in the live migration image. There
is no situation in which migration services and activation
is installed on the same host. As both packages are build
from the same python sources they have to conflict