Remove SMT references (#223)
SMT is no longer supported by SUSE and users are encouraged to migrate
to RMT. We want users to use RMT servers, or SCC as targets for receiving
system updates. Also improve requirements description and make the requirements
for the repository server more explicit w.r.t. availability of the proper channels.