Add an option to create a 'fat' initrd
This change adds a new systemd service, regenerate-initrd that is
responsible for running the dracut command when the config setting
'build_host_independent_initrd' is enabled.

This service is configured to run before the kernel-load service
and allows a new initrd to be created after the migration is complete
and before the final system reboot is performed.