Fixed reading of extra config file
The MigrationConfig() was read in only once in the mount
service. However for the container use case there is mount
service. Thus the config file information was missing.
For a more robust availability of the config information
this patch adds the reading to all services that consumes it.
In addition the change also prevents writing an 'all' status
from systemd to the log file. This information is not relevant
as the log contains information about the service runs. The data
also floods the log with a lot of useless information and
makes debugging more complicated