Add Container Workflow Feature
The current DMS workflow is based on a reboot either through
kexec or grub(loopsetup). The reboot is required to meet the
documented requirement for an offline upgrade process which
is the only supported process to upgrade from SLE12 to SLE15.
However this limitation is not a technology limitation, it
was based on product management decisions back in time.
From todays perspective it is possible to run a seamless
online upgrade from SLE12 to SLE15 using container technology.
This commit implements support for it in the DMS and provides the
required technology stack at:

https://build.opensuse.org/project/show/home:marcus.schaefer:dms

I'm aware that this container based workflow can only serve
as an additional/experimental workflow without guarantees to the
user as it would violate a PM decision. However, I also believe
that we have to provide good solutions when they solve a
number of severe issues that exists with the current DMS
reboot workflow. Along with this implementation I also prepared
a presentation to show and demo the workflow. Maybe this gains
some interest in the future and can help to start a PM
conversation about the SLE12 upgrade process.

In any way I'd like to add this additional workflow to the DMS
such that our users have a chance to upgrade when the reboot
way fails. If customers would consider it, it must be clear
to them that they are on unsupported territory. This information
must be placed to the documentation prior merge