Refactor packaging for container services
The current suse-migration-services package also packaged the
python code for the services. This is not compliant to python
refactoring in SUSE which allows multi-python version packages.
Because of this reason the package changed into a proper
python-migration package which build suse-migration-services
as a sub-package only containing the systemd service files.
python-migration itself builds for python 3.11 and therefore
also moves the migration code into the latest python version
supported on SLE15 onwards.