Add success indicator flag file
Added /var/log/distro_migration.exitcode to be written
by the migrate service. The file contains the exit code
from the zypper process and is used in the migrate tool
to know about the overall state of the migration process.
The migrate tool initializes the flag file with a failed
value.