Add package sources for suse-migration-rpm
The suse-migration-rpm package contains an OBS hook
script which turns the migration live ISO image into
a package named according to the name of the image
name attribute. In our case: SLES15-Migration. The
sources of this package was not maintained in any
git repository but are tightly coupled to the live
image build which resides here. Thus and to complete
the source gap we have for the DMS this commit adds
the sources for the meta package here