Increase version for SLES15-Migration package
The version of the SLES15-Migration package is taken from the
kiwi image description. Thus the bump of the version has
happened in this file. In addition a proper version constraint
to the suse-migration-sle15-activation package was added
due to the latest change from commit #1dbb055bc which now
installs the migration live ISO from the SLES15-Migration
package to /migration-image and no longer to /usr/share/migration-image