Fixed changes inconsistencies
Several changelog entries of the submitted package contains
bugzilla references which are not part of the git log.
Something that should never happen but well. This commit
updates the release tool to the latest version and adds
a new .changes and .changes.ref file within the inconsistencies
got fixed. In addition no longer existing e-mail addresses
will be replaced by their successor