Add product migration check
Add check which sends an API call to the registration server
migrations API to ask for the possible upgrade path. This is
acually the same call done by the zypper migration plugin but
prior the actual migration starts. The request can only be
issued if product and credentials data as well as the
registration server URL could be found on the system to
migrate. If the required information to issue the request
cannot be found, the check is skipped. In case the regsitration
server responds with an error, the pre-check prints this
information such that the user knows about the issue prior
starting the migration. The error information sent from the
registration server could also include instructions how to
fix the situation such that a migration can succeed. For
example on LTSS registered systems the request will respond
with an error message that products with activated LTSS
cannot be migrated and how to call SUSEConnect to disable
LTSS such that a migration will become possible.