Fix and cleanup update_bootloader service
The service tries to install the shim package by a chroot call
into the upgraded system. This is not going to work for several
reasons. First and foremost the call is interactive and was missing
several zypper options to allow this to be done non-interactive.
Second zypper from the DMS host should be called in the same way
as it was done during the actual migrate service for consistency
and stability reasons. When zypper has worked during the migrate
service it's likely to work for subsequent calls too. In addition
to the actual zypper call I reduced the code at several stages.
There is no need to catch an exception from Command.run to
raise the same exception (DistMigrationCommandException) that
would be done by Command.run itself. Also the command interface
logs the executed command. There is no reason to log the same
information again. This reduces the implementation by a number
of superfluous code and test code.