Relax on the shim-install call
The DMS tries to make the system gain secure boot capabilities.
However, several reasons e.g. system was not EFI based can
cause an error when attempting the setup. This action should
not cause the entire migration to fail. The DMS upgrades a
system based on the content of the existing disk layout.
If the disk layout to begin with is not suitable for EFI
or secure boot, the DMS will not interfere with the foundation
of the system during upgrade. However, in such a case the
DMS should still be able to upgrade the OS.