Fix suse-migration-console-log service
The console logging was based on forcing a dialog program on the
main console. If the main console also runs a getty which is by
default the case we run into several signal conflicts. This battle
is always won by the getty because it restarts always whereas the
console log service only restarts on failure. This leads to very
weird conditions such as no log displayed on the console at all
up to only a fraction of the log shown leading to wrong assumptions
on the current state of the migration. This commit now changes
the following:

1. The console log is forced into its own tty (8) and the
   service actively switches to this terminal

2. If the dialog ends (e.g user hits Exit) the service
   switches back to the first virtual terminal and restarts
   after 10 seconds. This is enough time to stop the
   log service in case it's unwanted, e.g. for debugging
   on this console

3. The dialog logging does no longer run on serial terminals.
   dialog does not really support to work well on serial
   consoles. The output is often completely broken and input
   sequences requires its own terminal. If the migration
   system is accessed from a terminal that doesn't show
   progress information a message in etc/motd has been
   added which provides information how to see progress