Drop obsolete check for resolv.conf
There is a file check for a resolv.conf file at a very early
stage of setup_host_network.py. This check shadows the init
log information and is also pointless because further on in
the code there is a proper handling for the resolv.conf
presence via has_host_resolv_setup(resolv_conf). This
Fixes #341