Fix product related requirement settings
The activation as well as the migration live image packages
set a specific product requirement for the product named
product(SLES). However this is in conflict with all products
SUSE creates that are based on SLES. One example here is
SLES for SAP which defines its own product product(SLES_SAP).
The activation of the migration should not be product specific.
These packages just sets up a bootloader setup to boot the
respective live image. Therefore this commit drops the
product requirement from the activation packages. The live
image however must be product specific and therefore this
commit adds code that allows to identify the target
product by the given image name and sets the respective
product requirement to the package that wraps the image.
This change also follows the new strategy to provide
migration live images for specific SUSE products, beginning
with SAP. This Fixes #349 and Fixes #354