Fix error evaluation from offline_migrations API
SCC provides the systems/products/offline_migrations API which
is used by the DMS to check if there is an upgrade path available.
In case of an existing upgrade path the response data type is
a list of dicts. In case of an error the response datatype is
just a dict. Checking for get() on a list type leads to an
exception reported as a problem with the upgrade path. This
commit fixes it.