Setup ExclusiveArch for activation packages
For migrations of SLE12 to SLE15 the DMS was used in
the cloud context only. grub based activation is supported
there only on x86_64 and aarch64

For migrations of SLE15 to SLE16 the DMS is used in a broader
context across SUSE products and architectures. grub based
activation is supported for all archs except s390. On s390
we support the kexec (run_migration) based method for
activation only.

This commit sets the ExclusiveArch option respectively