Follow up fix for the wicked to nm migration
The unit test to cover the copy of nm connection files
did not mock the glob call which leads to broken test
runs on systems that uses nm connections. This commit
fixes that part. The commit also fixes the way the cp
call is issued with the result of a glob call. In case
glob does not match the cp call would be done with only
one parameter and would cause a raise condition which
I believe is unwanted. If there are no nm connection
files the code should just not copy them. This commit
also changes that part along with the broken unit test.