store snapper pre snapshot in run_migration
The migration process is activated by either of two processes: the
installation of the activation package, or the run_migration
script. We store snapper pre snapshot in %pre in the first way, but
didn't do that when the migration is initialized by run_migration
before. This commit adds the missing piece.