Add changes for newest wicked2nm
With the new wicked2nm version 1.3.0 hopefully wicked2nm-continue-migration is
less necessary so this removes the default true in SLE16.
Also there are some changes to the netconfig handling so it makes more sense to
read the base netconfig dir instead of copying the files to
'/var/cache/wicked_config'.
I've also added an explanation to the pre-checks to explain how to apply the
`--continue-migration` flag to the migration.