Create systemd.link files for virtual VMware/Hyper-V NICs
The 75-persistent-net-generator.rules does not create udev-rules
for virtual nics, matching the hwaddr "00:0c:29:*|00:50:56:*" (VMware)
or "00:15:5d:*" (Hyper-V).
When migrating to SLE16, predictable ifnames are enabled and it
is very likely, that the nic will get a different name.

To fix this, create 'systemd.link' files matching the DRIVER and
PCI_ID to ensure the virtual NIC retains its  expected name.
(bsc#1250076)