Fixed selinux to apparmor migration
The selinux to apparmor migration is based on a grub default
setting change followed by the installation of a pattern package.
It's important that this action is done prior the network migration
from wicked to NetworkManager because this migration can lead
to network inconsistencies which prevents the installation of
packages from the remote repos. So happened in cloud instances
and also in the container based workflow. As such this commit
makes sure the selinux to apparmor migration happens before
the network migration. In addition the selinux to apparmor
migration should not mark the entire migration as failed as
such a problem can also be fixed in the migrated system later