Fixed behavior of wicked2m pre check
The output produced by the pre-check is in my opinion very
misleading. It outputs information from wicked2m which says
that warnings can be ignored by passing an option to the
tool but for the actual DMS system migration this has no
meaning. The additional information to place a config
setting to a DMS config file is correct but completely
confusing with regards to the first information provided.
This commit refactors the code which also doesn't use
the existing command interface and imho can be done
simpler. In addition the precheck runs code when inside
of the migration, but this code is useless because the
real (no dry-run) call of wicked2nm happens in the
setup_host_network service.