Fix MinSLEVersion value depending on target
The MinSLEVersion global variable was set to a value
suitable for migrations from SLE12 to SLE15. Now that the
scope of the DMS is not only this target we have to add
a mechanism to dynamically set an appropriate value.
I added explanatory comments which hopefully makes sense
to explain when this information is taken into account