Fixed LSM pre checks to be more robust
When called on systemd without aa-status installed, the code
runs into a traceback. Looking at the logic flow I believe this
can be done more robust by directly checking for the availabilty
of the tool. In addition the code uses python type hints which
do not exists on the python used in SLE12 for which the code needs
to stay compatible for some time.