Simplify interface naming by disabling predictable names at boot
Currently our most difficult problem with regards to the network interface
naming scheme is that when the migration image boots the interfaces get
predictable names like 'enp1s0' which we then have to revert/write custom rules
around. To simplify this we can install a file at
/etc/systemd/network/99-default.link. This skips all naming schemes and the
kernel names are used. Then we can remove this file again and copy over all the
necessary udev rules and link files from the host and with udevadm trigger apply
these. This should imitate the behavior of the final migrated system most
closely.