Simplify interface naming by booting with net.ifnames=0
Currently our most difficult problem with regards to the network interface
naming scheme is that when the migration image boots the interfaces get
predictable names like 'enp1s0' which we then have to revert/write custom rules
around. To simplify this we can boot with the net.ifnames=0 applied. This skips
all naming schemes and the kernel names are used. Then we can copy over all the
necessary udev rules and link files from the host and with udevadm trigger apply
these. This is only necessary in the migration image because these files can't
be present on the system at boot. On the migrated system since these files are
already present, the kernel commandline parameter isn't necessary anymore
because udev rules and link files take priority over the predictable naming
scheme.