Fixed azuremetadata device lookup
On Azure the azuremetadata call is prepared with the --device
option pointing to the disk containing the system. This disk
device is obtained via the get_root_disk_device method which
makes use of the findmnt tool to read the device that contains
the root filesystem mountpoint. If that filesystem is using
sophisticated technology like subvolumes on btrfs, findmnt
reports this information in addition to the actual unix block
device in brackets [...]. For the logic flow of get_root_disk_device
this extra information broke the subsequent lsblk call. This
commit strips eventually existing extra block device information
from the findmnt call output.